requirements.txt
#python==3.11
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scikit-image==0.23.2
tqdm==4.66.2
tensorboard==2.16.2
torch_geometric==2.5.3
opencv-python==4.9.0.80
PyYAML==6.0.1
xgboost==2.1.1
networkx==3.4.2
scipy==1.13.0
cellpose==3.0.8
openslide-python==1.2.0